+++ src/settings/pyproject.rs
@@ -131,7 +131,8 @@ mod tests {
 
     use crate::registry::RuleCodePrefix;
     use crate::rules::flake8_quotes::settings::Quote;
-    use crate::rules::flake8_tidy_imports::settings::{ApiBan, Strictness};
+    use crate::rules::flake8_tidy_imports::banned_api::ApiBan;
+    use crate::rules::flake8_tidy_imports::relative_imports::Strictness;
     use crate::rules::{
         flake8_bugbear, flake8_errmsg, flake8_import_conventions, flake8_pytest_style,
         flake8_quotes, flake8_tidy_imports, mccabe, pep8_naming,
